Added OC optimizer
The Optimality Criterion method [OC] is a simple heuristic algorithm to
perform the update of the design variables. This method is one of the
earlier methods used to solve topology optimization problems and only
allow for imposition of a single constraint.

Today the Method of Moving Asymptotes [MMA], see svanberg, is the most
widely used algorithm for topology optimization. One of the advantages
of the MMA is that it allows for more flexibility than the OC method.

bla.bla.bla :)